+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `Release::asset_for` now searches for current `OS` and `ARCH` inside `asset.name` if `target` failed to match
 - Update `reqwest` to `0.12.0`
 - Update `hyper` to `1.2.0`
+- Replace `zip` with `zip_next`
 ### Removed
 
 ## [0.39.0]